requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
more-itertools==10.8.0
numpy==2.3.5
packaging==25.0
pillow==12.0.0
premailer==3.10.0
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
